モデルの分割
